Don't treat Z_BUF_ERROR as a fatal error
Previously Z_BUF_ERROR error code was treated as fatal and it raised an exception in OCaml. However the error is not fatal according to zlib documentation [1].

Handling this is necessary for streaming compression where you never finish the stream with Z_FINISH.

Seems like ignoring the error is sufficient as returned values for used_in and used_out will be 0 in case it happens.

Documentation says:
"this is simply an indication that deflate() could not consume more input or produce more output. deflate() can be called again with more output space or more available input, which it will be in this code."

[1] https://zlib.net/zlib_how.html