Full support ZIP64 archive format
This adds support for individual files larger than 4 Gib and for archives containing more than 65535 files, both for reading and for writing.

When adding files to ZIP archive, stop adding a data descriptor after the file data, as it is unclear how to do this correctly for files larger than 4 Gib. Instead, patch the local file header with the final sizes and CRC.

Closes: #35