Zlib.inflate: detect truncated input instead of looping
Truncated input can cause `inflate` to produce no output and to return
"not finished" infinitely.

Detect this case and fail cleanly with an exception.

Note: it might be enough to fail at the first time `inflate` returns
"not finished but zero output produced".  Here we retry a few times
before failing, just in case.